Change createGroups() in AbstractNetworkClusterer to public static
